- stockService -> employeeService - @ApiOperation 코멘트 정리 - 직원 인터페이스 기능 추가(주문 목록 조회, 주문 상태 변경) - 주문 5회 이상 시, 10% 할인을 위해 totalPrice의 형식을 double로 수정 - 주문 상태에 completed(배달 완료) 추가 - userService에서 saveUser 메서드의 반환형을 void로 변경

+++ src/main/java/com/sogong/tejava/entity/Order.java
@@ -15,7 +15,7 @@ public class Order extends BaseTimeEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private int totalPrice;
+    private double totalPrice;
     private OrderStatus orderStatus;
     private String optionToString;
